Re-organise classes to not expose Renderer Implementation
