+++ lib/src/web/rtc_peerconnection_factory.dart
@@ -2,18 +2,17 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:html' as html;
 
-import 'package:flutter_webrtc/src/interface/navigator.dart';
-import 'package:flutter_webrtc/src/web/navigator_impl.dart';
-
 import '../interface/factory.dart';
 import '../interface/media_recorder.dart';
 import '../interface/media_stream.dart';
+import '../interface/navigator.dart';
 import '../interface/rtc_peerconnection.dart';
 import '../interface/rtc_video_renderer.dart';
 import 'media_recorder_impl.dart';
 import 'media_stream_impl.dart';
+import 'navigator_impl.dart';
 import 'rtc_peerconnection_impl.dart';
-import 'rtc_video_view_impl.dart';
+import 'rtc_video_renderer_impl.dart';
 
 Future<RTCPeerConnection> createPeerConnection(
     Map<String, dynamic> configuration,
